backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.4.0
python-multipart==0.0.9
opencv-python-headless==4.10.0.84

# TensorFlow per OS (avoid tensorflow-cpu 2.15 on macOS/Python 3.12)
tensorflow-cpu==2.17.0; platform_system!="Darwin"
tensorflow-macos==2.16.1; platform_system=="Darwin"
tensorflow-metal==1.2.0; platform_system=="Darwin"

# NumPy compatible with TF 2.16/2.17
numpy>=1.26.4,<2.1